+++ examples/snippets/app/examples/feature-flags-in-proxy/flags.ts
@@ -1,0 +1,8 @@
+import { flag } from 'flags/next';
+
+export const basicProxyFlag = flag<boolean>({
+  key: 'basic-proxy-flag',
+  decide({ cookies }) {
+    return cookies.get('basic-proxy-flag')?.value === '1';
+  },
+});