+++ examples/snippets/app/examples/feature-flags-in-proxy/handlers.ts
@@ -1,17 +1,17 @@
 'use client';
 
 export function actAsFlaggedInUser() {
-  document.cookie = 'basic-edge-middleware-flag=1; Path=/';
+  document.cookie = 'basic-proxy-flag=1; Path=/';
   window.location.reload();
 }
 
 export function actAsFlaggedOutUser() {
-  document.cookie = 'basic-edge-middleware-flag=0; Path=/';
+  document.cookie = 'basic-proxy-flag=0; Path=/';
   window.location.reload();
 }
 
 export function clear() {
   document.cookie =
-    'basic-edge-middleware-flag=; Path=/; Expires=Thu, 01 Jan 1970 00:00:00 GMT';
+    'basic-proxy-flag=; Path=/; Expires=Thu, 01 Jan 1970 00:00:00 GMT';
   window.location.reload();
 }