+++ packages/adapter-vercel/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @flags-sdk/vercel
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies [620974c]
+  - @vercel/flags-core@0.1.8
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [43293a3]
+  - @vercel/flags-core@0.1.7
+
 ## 0.1.6
 
 ### Patch Changes